Switching commands to i18n.
Going through them one by one and finishing the build on en-US.json

I think this takes care of the fun, geocaching, and info groups for now and some of admin.  Need to do the rest of admin and chat the move on to prefix commands, events, and functions probably... ...but... ...I am done for tonight.